requirements.txt
openai==0.28
marvin
requests
beautifulsoup4
google-cloud-storage